requirements.txt
reportlab~=4.2.0
pandas~=2.2.2
PyPDF2~=3.0.1
gspread~=6.1.2
oauth2client~=4.1.3
google~=3.0.0
openpyxl~=3.1.4
requests~=2.32.3
pillow~=10.3.0
selenium~=4.24.0
webdriver-manager~=4.0.2
dropbox~=12.0.2
python-dotenv~=1.0.1
psycopg2~=2.9.9
google-auth-oauthlib~=1.2.1
ImageHash~=4.3.1
ngrok~=1.4.0
Flask~=3.0.3
setuptools~=75.1.0
protobuf~=5.28.2
beautifulsoup4~=4.13.0b2
google-api-python-client~=2.149.0
pyTelegramBotAPI~=4.24.0